labs/azure-agentic-ai-architect-labs/09-end-to-end-reference-implementation/requirements.txt
# Runtime dependencies
aiohttp>=3.8.0
aiofiles>=23.0.0

# Azure SDK
azure-identity>=1.14.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.17.0
azure-cosmos>=4.4.0

# OpenAI and LLM integrations
openai>=1.3.0
semantic-kernel>=0.3.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.24.0
ruff>=0.3.0

# Monitoring and logging
python-json-logger>=2.0.0
prometheus-client>=0.18.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
loguru>=0.7.2
